backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
daphne
Pillow
python-dotenv
